refactor WalletNetworkFacade
