+++ fearless/Common/Network/WalletNetworkFacade+Protocol.swift
@@ -111,25 +111,6 @@ extension WalletNetworkFacade: WalletNetworkOperationFactoryProtocol, RuntimeCon
         )
     }
 
-    private func fetchMinimalBalanceOperation() -> CompoundOperationWrapper<BigUInt> {
-        let codingFactoryOperation = runtimeCodingService.fetchCoderFactoryOperation()
-        let constOperation = PrimitiveConstantOperation<BigUInt>(path: .existentialDeposit)
-        constOperation.configurationBlock = {
-            do {
-                constOperation.codingFactory = try codingFactoryOperation.extractNoCancellableResultData()
-            } catch {
-                constOperation.result = .failure(error)
-            }
-        }
-
-        constOperation.addDependency(codingFactoryOperation)
-
-        return CompoundOperationWrapper(
-            targetOperation: constOperation,
-            dependencies: [codingFactoryOperation]
-        )
-    }
-
     func fetchTransactionHistoryOperation(
         _ request: WalletHistoryRequest,
         pagination: Pagination
